scripts/imagefile.py
from lancedb.embeddings import EmbeddingFunctionRegistry
from lancedb.pydantic import LanceModel, Vector
from PIL import Image

registry = EmbeddingFunctionRegistry.get_instance()
clip = registry.get("open-clip").create()

class ImageFile(LanceModel):
    vector: Vector(clip.ndims()) = clip.VectorField()
    image_uri: str = clip.SourceField()

    @property
    def image(self):
        return Image.open(self.image_uri)