requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
lancedb==0.13.0
pandas==2.2.2
open-clip-torch==2.26.1
pillow==10.4.0